+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Prowide ISO 20022 - CHANGELOG
 
-### 10.2.10 - SNAPSHOT
+### 10.2.10 - November 2025
   * (PW-2731) Enhanced camt.057 amount extraction to fallback to Itm/Amt when TtlAmt is not present
   * (GH-174) Fix: MinimumEscapeHandler duplicating quote characters in non-attribute contexts
   * (GH-168) Fix: MX parsing error when AppHdr appears before Document element in xsys messages